dev-requirements.txt
aiohttp==3.8.4
aiologger==0.7.0
aiosignal==1.3.1
alembic==1.9.3
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.91.0
fastapi-pagination==0.11.4
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
motor==3.1.1
multidict==6.0.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.4
pyflakes==3.0.1
PyJWT==2.5.0
pymongo==4.3.3
python-dotenv==0.21.1
python-jose==3.3.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.24.0
typing_extensions==4.4.0
uvicorn==0.20.0
websockets==10.4
yarl==1.8.2
